Update the feedback & typo
